refactor(test): move Transport constant tests to transport_test.go
TestMessageTypeConstants and TestStatusCodeConstants test types defined in
transport.go; frame_test.go focuses on Frame and sentinel error behaviour.
Moving them to a dedicated transport_test.go improves discoverability.

Also narrows the CHANGELOG [Unreleased] Added entry for MessageType to
focus on the new type only — TextMessage/BinaryMessage already existed and
are described under Changed.

Fixes Copilot review comments on PR #28.
